tests/submission/MOSQUITO/event-source/eventsource-aborted-after-close-method.html

<!DOCTYPE html>
<html>
<head>
	<title>The close() method must abort any instances of the fetch algorithm started for this EventSource object, and must set the readyState attribute to CLOSED:</title>
  <meta charset="utf-8" />
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/resources/WebIDLParser.js"></script>
	<script src="/resources/idlharness.js"></script>
</head>
<body>
	<h1>The close() method must abort any instances of the fetch algorithm started for this EventSource object, and must set the readyState attribute to CLOSED:</h1>
	<div id="log"></div>
  <script>
		setup({timeout:10000});
		var source = new EventSource('http://html5.mosquito-fp7.eu/messages/t0022');
		var tclose = async_test("The close() method must set the readyState attribute to CLOSED (2)");
    
    source.addEventListener('open', function(e){ 
    	/* Connection opened. */
    }, false);
    	
    source.addEventListener('message', function(e){
 	    	source.close();
			tclose.step(function() {assert_equals(source.readyState, EventSource.CLOSED)});
			tclose.done();
    }, false);
    
    source.addEventListener('error', function(e){
    	/* Connection closed. */
     }, false);
    
  </script>
  </div>
</body>
</html>
